refactor(NavBar): use `ExternalLink` component to simplify discord link
Rather than having NavbarItem need to distinguish between a webapp or electron version, let's wrap the item in an `ExternalLink` component and have that component control the linking behavior
