--- app/components/nav/platformSpecific/NavSidebar.electron.ts
@@ -1,5 +1,0 @@
-import { shell } from 'electron'
-
-export function openInExternalWindow (uri: string) {
-  shell.openExternal(uri)
-}
